Added Checkboxes

+++ dir_copy.py
@@ -2,16 +2,25 @@ import pathlib
 import shutil
 import sys
 from typing import List
+from enum import IntEnum
+
+
+class Select(IntEnum):
+    DIR = 1
+    FILE = DIR << 1
+    ALL = (DIR << 2) - 1
 
 
 class DirSync:
-    def __init__(self, src: pathlib.Path, dst: pathlib.Path, ignore: List[str] = None, verbose=False):
+    def __init__(self, src: pathlib.Path, dst: pathlib.Path, ignore: List[str] = None,
+                 selected: Select = Select.ALL, verbose=False):
         assert src.exists(), "Source does not exist."
         assert dst.exists(), "Destination does not exist."
 
         self.src = src
         self.dst = dst
         self.ignore = list()
+        self.selected = selected
         if ignore is not None:
             self.ignore = ignore
         self.verbose = verbose
@@ -26,14 +35,14 @@ class DirSync:
         for dir_child in src.iterdir():
             if dir_child.name in self.ignore:
                 continue
-            if dir_child.is_dir():
+            if dir_child.is_dir() and self.selected & Select.DIR:
                 if not (dst / dir_child.name).exists():
                     sys.stderr.write("Copy " + str(dir_child) + '\n')
                     shutil.copytree((src / dir_child.name), (dst / dir_child.name))
                 else:
                     dir_rel_child = (dir_rel / dir_child).relative_to(self.src)
                     self.sync_dir(dir_rel_child)
-            elif dir_child.is_file():
+            elif dir_child.is_file() and self.selected & Select.FILE:
                 if not (dst / dir_child.name).exists():
                     sys.stderr.write("Copy " + str(dir_child) + '\n')
                     shutil.copy((src / dir_child.name), (dst / dir_child.name))
@@ -42,13 +51,13 @@ class DirSync:
         for item in self.src.iterdir():
             if item.name in self.ignore:
                 continue
-            if item.is_dir():
+            if item.is_dir() and self.selected & Select.DIR:
                 if not (self.dst / item.name).exists():
                     shutil.copytree((self.src / item.name), (self.dst / item.name))
                 else:
                     dir_rel = item.relative_to(self.src)
                     self.sync_dir(dir_rel)
-            elif item.is_file():
+            elif item.is_file() and self.selected & Select.FILE:
                 shutil.copy((self.src / item.name), (self.dst / item.name))
         print("Done!")
